[FORM] Adição de campo tipo data. Parte de máscara e "datepicker".

+++ form/core/src/main/java/br/net/mirante/singular/form/mform/util/comuns/MTipoAnoMes.java
@@ -24,7 +24,7 @@ public class MTipoAnoMes extends MTipoSimples<MIAnoMes, YearMonth> {
     @Override
     public YearMonth converterNaoNativoNaoString(Object valor) {
         if (valor instanceof Integer) {
-            return converterFromInteger(((Integer) valor).intValue());
+            return converterFromInteger((Integer) valor);
         } else if (valor instanceof Date) {
             Calendar cal = new GregorianCalendar();
             cal.setTime((Date) valor);
